+++ src/config/settings.ts
@@ -44,6 +44,7 @@ const settings: Settings = {
     blockSyncBatchSize: parseInt(process.env.BLOCK_SYNC_BATCH_SIZE || '5', 10),
     feedsFile: process.env.FEEDS_FILE || '',
     feedsOnChain: process.env.FEEDS_ON_CHAIN_FILE || '',
+    layer1FeedFile: process.env.LAYER1_FEEDS_FILE || '',
   },
   blockchain: {
     contracts: {
@@ -58,6 +59,44 @@ const settings: Settings = {
     homeChain: {
       chainId: 'bsc',
       replicationConfirmations: parseInt(process.env.HOME_REPLICATION_CONFIRMATIONS || '20', 10),
+    },
+    blockchainScanner: {
+      arbitrum: {
+        startBlockNumber: parseInt(process.env.ARBITRUM_SCANNER_START_BLOCK_NUMBER || '0', 10),
+        scanBatchSize: parseInt(process.env.ARBITRUM_BLOCK_SCAN_BATCH_SIZE || '1000', 10),
+        fetchBlocksBatchSize: parseInt(process.env.ARBITRUM_SCANNER_FETCH_BLOCKS_BATCH || '100', 10),
+        maxRequestConcurrency: parseInt(process.env.ARBITRUM_MAX_REQUEST_CONCURRENCY || '10', 10),
+        confirmations: parseInt(process.env.ARBITRUM_BLOCK_CONFIRMATIONS || '20', 10),
+        providerUrl: process.env.ARBITRUM_SCANNER_PROVIDER_URL, // we can't have default providers set up
+        contractRegistryAddress: process.env.ARBITRUM_REGISTRY_CONTRACT_ADDRESS,
+      },
+      base: {
+        startBlockNumber: parseInt(process.env.BASE_SCANNER_START_BLOCK_NUMBER || '0', 10),
+        scanBatchSize: parseInt(process.env.BASE_BLOCK_SCAN_BATCH_SIZE || '1000', 10),
+        fetchBlocksBatchSize: parseInt(process.env.BASE_SCANNER_FETCH_BLOCKS_BATCH || '100', 10),
+        maxRequestConcurrency: parseInt(process.env.BASE_MAX_REQUEST_CONCURRENCY || '10', 10),
+        confirmations: parseInt(process.env.BASE_BLOCK_CONFIRMATIONS || '10', 10),
+        providerUrl: process.env.BASE_SCANNER_PROVIDER_URL, // we can't have default providers set up
+        contractRegistryAddress: process.env.BASE_REGISTRY_CONTRACT_ADDRESS,
+      },
+      linea: {
+        startBlockNumber: parseInt(process.env.LINEA_SCANNER_START_BLOCK_NUMBER || '0', 10),
+        scanBatchSize: parseInt(process.env.LINEA_BLOCK_SCAN_BATCH_SIZE || '1000', 10),
+        fetchBlocksBatchSize: parseInt(process.env.LINEA_SCANNER_FETCH_BLOCKS_BATCH || '100', 10),
+        maxRequestConcurrency: parseInt(process.env.LINEA_MAX_REQUEST_CONCURRENCY || '10', 10),
+        confirmations: parseInt(process.env.LINEA_BLOCK_CONFIRMATIONS || '10', 10),
+        providerUrl: process.env.LINEA_SCANNER_PROVIDER_URL, // we can't have default providers set up
+        contractRegistryAddress: process.env.LINEA_REGISTRY_CONTRACT_ADDRESS,
+      },
+      polygon: {
+        startBlockNumber: parseInt(process.env.POLYGON_SCANNER_START_BLOCK_NUMBER || '0', 10),
+        scanBatchSize: parseInt(process.env.POLYGON_BLOCK_SCAN_BATCH_SIZE || '1000', 10),
+        fetchBlocksBatchSize: parseInt(process.env.POLYGON_SCANNER_FETCH_BLOCKS_BATCH || '100', 10),
+        maxRequestConcurrency: parseInt(process.env.POLYGON_MAX_REQUEST_CONCURRENCY || '10', 10),
+        confirmations: parseInt(process.env.POLYGON_BLOCK_CONFIRMATIONS || '15', 10),
+        providerUrl: process.env.POLYGON_SCANNER_PROVIDER_URL, // we can't have default providers set up
+        contractRegistryAddress: process.env.POLYGON_REGISTRY_CONTRACT_ADDRESS,
+      },
     },
     multiChains: {
       bsc: {
@@ -162,7 +201,7 @@ const settings: Settings = {
     maxSearchRange: parseInt(process.env.MAX_SIGNATURE_SEARCH_RANGE || '8', 10),
     maxSearchRangeInSeconds: parseInt(process.env.MAX_SIGNATURE_SEARCH_RANGE || '8', 10) * 24 * 60 * 60,
   },
-  name: process.env.NEW_RELIC_APP_NAME || process.env.NAME || 'default',
+  name: process.env.NAME || 'default',
 };
 
 export default settings;