+++ src/config/setupNewrelic.ts
@@ -1,2 +1,0 @@
-process.env['NEW_RELIC_ENABLED'] ||= 'false';
-import 'newrelic';