+++ src/services/ForeignChainReplicator.ts
@@ -1,6 +1,5 @@
 import { inject, injectable } from 'inversify';
 import { Logger } from 'winston';
-import newrelic from 'newrelic';
 
 import { BlockChainDataFactory } from '../factories/BlockChainDataFactory';
 import { IForeignBlockReplicator, SolanaBlockReplicator } from './foreign-chain';
@@ -134,7 +133,7 @@ export class ForeignChainReplicator {
   ) => {
     const { errorLimit, warningLimit } = this.settings.blockchain.multiChains[chainId].transactions.mintBalance;
 
-    if (balance.lt(toCurrency(errorLimit))) {
+    if (!balance || balance.lt(toCurrency(errorLimit))) {
       throw new Error(`[${chainId}] Balance (${address.slice(0, 10)}) is lower than ${errorLimit}`);
     }
 
@@ -144,7 +143,6 @@ export class ForeignChainReplicator {
   };
 
   private noticeError = (err: string): void => {
-    newrelic.noticeError(Error(err));
     this.logger.error(err);
   };
 }